requirements.txt
viam-sdk==0.28
python-dotenv
  grpclib~=0.4.6
  googleapis-common-protos~=1.60.0
  typing-extensions~=4.8.0
  protobuf~=4.24.3
